chore: reorganise structure of GUI code.
