+++ src/gui/widgets/dialogs/view_input_dialog.py
