+++ src/gui/widgets/slots/beam_slots.py
