+++ src/gui/widgets/slots/instrument_slots.py
