+++ src/gui/widgets/slots/sample_background_slots.py
