+++ src/gui/widgets/tables/beam_profile_table.py
@@ -1,0 +1,197 @@
+from PySide6.QtCore import Qt
+from PySide6.QtWidgets import QTableView, QDoubleSpinBox
+from src.gui.widgets.tables.gudpy_tables import GudPyTableModel, GudPyDelegate
+
+class BeamProfileModel(GudPyTableModel):
+    """
+    Class to represent a BeamProfileModel. Inherits GudPyTableModel.
+
+    ...
+
+    Methods
+    -------
+    columnCount(parent)
+        Returns the number of columns in the model.
+    headerData(section, orientation, role)
+        Stub method.
+    setData(index, value, role)
+        Sets data in the model.
+    data(index, role)
+        Returns data at a given index
+    insertRow(data)
+        Inserts a row of data into the model.
+    """
+    def __init__(self, data, headers, parent):
+        """
+        Calls super().__init__ on the passed parameters.
+        Parameters
+        ----------
+        data : list
+            Data for model to use.
+        headers: str[]
+            Column headers for table.
+        parent : QWidget
+            Parent widget.
+        """
+        super(BeamProfileModel, self).__init__(data, headers, parent)
+
+    def columnCount(self, parent):
+        """
+        Returns the number of columns in the model.
+        Parameters
+        ----------
+        parent : QWidget
+            Parent widget.
+        Returns
+        -------
+        int
+            Number of columns in the model - this is always 1.
+        """
+        return 1
+
+    def headerData(self, section, orientation, role):
+        """
+        There are no headers, so this always returns nothing.
+        Returns
+        -------
+        None
+        """
+        pass
+
+    def setData(self, index, value, role):
+        """
+        Sets data in the model.
+        Parameters
+        ----------
+        index : QModelIndex
+            Index to set data at.
+        value : any
+            Value to set data to.
+        role : int
+            Role.
+        """
+        row = index.row()
+        if role == Qt.EditRole:
+            self._data[row] = value
+
+    def data(self, index, role):
+        """
+        Returns the data at a given index.
+        Parameters
+        ----------
+        index : QModelIndex
+            Index to return data from.
+        role : int
+            Role
+        Returns
+        -------
+        float
+            Data at given index.
+        """
+        row = index.row()
+        if role == role & (Qt.DisplayRole | Qt.EditRole):
+            return self._data[row]
+
+    def insertRow(self):
+        """
+        Inserts a row of data into the model.
+        The data is always by default 0.0.
+        Calls super().insertRow().
+        """
+        return super(BeamProfileModel, self).insertRow(0.)
+
+
+class BeamProfileDelegate(GudPyDelegate):
+    """
+    Class to represent a GroupingParameterDelegate. Inherits GudPyDelegate.
+
+    ...
+
+    Methods
+    -------
+    createEditor(parent, option, index)
+        Creates an editor.
+    """
+
+    def createEditor(self, parent, option, index):
+        """
+        Creates an editor, and returns it.
+        Parameters
+        ----------
+        parent : QWidget
+            Parent widget.
+        option : QStyleOptionViewItem
+            Option.
+        index : QModelIndex
+            Index in to create editor at.
+        Returns
+        -------
+        QDoubleSpinBox
+            The created editor.
+        """
+        editor = QDoubleSpinBox(parent)
+        editor.setMinimum(0)
+        editor.setMaximum(1)
+        editor.setSingleStep(0.01)
+        return editor
+
+
+class BeamProfileTable(QTableView):
+    """
+    Class to represent a BeamProfileTable. Inherits QTableView.
+
+    ...
+    Attributes
+    ----------
+    parent : QWidget
+        Parent widget.
+    Methods
+    -------
+    makeModel(data)
+        Creates the model using the data.
+    insertRow()
+        Inserts a row into the model.
+    removeRow(rows)
+        Removes selected rows from the model.
+    """
+    def __init__(self, parent):
+        """
+        Constructs all the necessary attributes
+        for the BeamProfileTable object.
+        Calls super().__init__
+        which calls the dunder init method from GudPyTableModel.
+        Parameters
+        ----------
+        parent : QWidget
+            Parent widget.
+        """
+        self.parent = parent
+        super(BeamProfileTable, self).__init__(parent=parent)
+
+    def makeModel(self, data):
+        """
+        Makes the model and the delegate based on the data.
+        Parameters
+        ----------
+        data : list
+            Data for model to use.
+        """
+        self.setModel(BeamProfileModel(data, [], self.parent))
+        self.setItemDelegate(BeamProfileDelegate())
+
+    def insertRow(self):
+        """
+        Inserts a row into the model.
+        """
+        self.model().insertRow()
+
+    def removeRow(self, rows):
+        """
+        Removes rows from the model.
+        Parameters
+        ----------
+        rows : QModelIndexList
+            Rows to be removed.
+        """
+        for _row in rows:
+            self.model().removeRow(_row.row())
